Refactor: 마이페이지 폴더 정리 및 모달 파일 확장자 수정
