+++ src/pages/MyPage.tsx
@@ -1,4 +1,4 @@
-import EditInfo from '@/components/EditInfo';
+import EditInfo from '@/components/myPage/EditInfo';
 
 const MyPage = () => {
   return (
